+++ microsite/sidebars.ts
@@ -228,6 +228,7 @@ export default {
           'features/software-templates/migrating-from-v1beta2-to-v1beta3',
           'features/software-templates/dry-run-testing',
           'features/software-templates/experimental',
+          'features/software-templates/template-extensions',
         ],
       },
       {