+++ packages/canon/src/components/Button/Button.stories.tsx
@@ -19,6 +19,7 @@ import { Button } from './Button';
 import { Flex } from '../Flex';
 import { Text } from '../Text';
 import { ButtonProps } from './types';
+import { Icon } from '../Icon';
 
 const meta = {
   title: 'Components/Button',
@@ -55,10 +56,10 @@ export const Variants: Story = {
   },
   render: () => (
     <Flex align="center">
-      <Button iconStart="cloud" variant="primary">
+      <Button iconStart={<Icon name="cloud" />} variant="primary">
         Button
       </Button>
-      <Button iconStart="cloud" variant="secondary">
+      <Button iconStart={<Icon name="cloud" />} variant="secondary">
         Button
       </Button>
     </Flex>
@@ -71,8 +72,12 @@ export const Sizes: Story = {
   },
   render: () => (
     <Flex align="center">
-      <Button size="medium">Medium</Button>
-      <Button size="small">Small</Button>
+      <Button size="medium" iconStart={<Icon name="cloud" />}>
+        Medium
+      </Button>
+      <Button size="small" iconStart={<Icon name="cloud" />}>
+        Small
+      </Button>
     </Flex>
   ),
 };
@@ -83,9 +88,13 @@ export const WithIcons: Story = {
   },
   render: args => (
     <Flex align="center">
-      <Button {...args} iconStart="cloud" />
-      <Button {...args} iconEnd="chevron-right" />
-      <Button {...args} iconStart="cloud" iconEnd="chevron-right" />
+      <Button {...args} iconStart={<Icon name="cloud" />} />
+      <Button {...args} iconEnd={<Icon name="chevron-right" />} />
+      <Button
+        {...args}
+        iconStart={<Icon name="cloud" />}
+        iconEnd={<Icon name="chevron-right" />}
+      />
     </Flex>
   ),
 };
@@ -96,9 +105,13 @@ export const FullWidth: Story = {
   },
   render: args => (
     <Flex direction="column" gap="4" style={{ width: '300px' }}>
-      <Button {...args} iconStart="cloud" />
-      <Button {...args} iconEnd="chevron-right" />
-      <Button {...args} iconStart="cloud" iconEnd="chevron-right" />
+      <Button {...args} iconStart={<Icon name="cloud" />} />
+      <Button {...args} iconEnd={<Icon name="chevron-right" />} />
+      <Button
+        {...args}
+        iconStart={<Icon name="cloud" />}
+        iconEnd={<Icon name="chevron-right" />}
+      />
     </Flex>
   ),
 };
@@ -150,22 +163,22 @@ export const Playground: Story = {
                 Button
               </Button>
               <Button
-                iconStart="cloud"
+                iconStart={<Icon name="cloud" />}
                 variant={variant as ButtonProps['variant']}
                 size={size as ButtonProps['size']}
               >
                 Button
               </Button>
               <Button
-                iconEnd="chevron-right"
+                iconEnd={<Icon name="chevron-right" />}
                 variant={variant as ButtonProps['variant']}
                 size={size as ButtonProps['size']}
               >
                 Button
               </Button>
               <Button
-                iconStart="cloud"
-                iconEnd="chevron-right"
+                iconStart={<Icon name="cloud" />}
+                iconEnd={<Icon name="chevron-right" />}
                 style={{ width: '200px' }}
                 variant={variant as ButtonProps['variant']}
                 size={size as ButtonProps['size']}
@@ -180,7 +193,7 @@ export const Playground: Story = {
                 Button
               </Button>
               <Button
-                iconStart="cloud"
+                iconStart={<Icon name="cloud" />}
                 variant={variant as ButtonProps['variant']}
                 size={size as ButtonProps['size']}
                 disabled
@@ -188,7 +201,7 @@ export const Playground: Story = {
                 Button
               </Button>
               <Button
-                iconEnd="chevron-right"
+                iconEnd={<Icon name="chevron-right" />}
                 variant={variant as ButtonProps['variant']}
                 size={size as ButtonProps['size']}
                 disabled