+++ packages/frontend-plugin-api/src/wiring/createFrontendFeatureLoader.test.ts
@@ -60,7 +60,7 @@ describe('createFrontendFeatureLoader', () => {
         );
         return [
           createFrontendPlugin({
-            id: `${pluginIdPrefix}-1`,
+            pluginId: `${pluginIdPrefix}-1`,
             extensions: [
               createExtension({
                 name: '1',
@@ -76,7 +76,7 @@ describe('createFrontendFeatureLoader', () => {
             ],
           }) as FrontendFeature | FrontendFeatureLoader,
           createFrontendPlugin({
-            id: `${pluginIdPrefix}-2`,
+            pluginId: `${pluginIdPrefix}-2`,
             extensions: [
               createExtension({
                 name: '2',
@@ -92,7 +92,7 @@ describe('createFrontendFeatureLoader', () => {
             ],
           }) as FrontendFeature | FrontendFeatureLoader,
           createFrontendPlugin({
-            id: `${pluginIdPrefix}-output`,
+            pluginId: `${pluginIdPrefix}-output`,
             extensions: [
               createExtension({
                 name: 'output',
@@ -166,7 +166,7 @@ describe('createFrontendFeatureLoader', () => {
         async loader(_) {
           return [
             createFrontendPlugin({
-              id: 'plugin-0',
+              pluginId: 'plugin-0',
               extensions: [
                 createExtension({
                   name: '0',
@@ -184,7 +184,7 @@ describe('createFrontendFeatureLoader', () => {
             createFrontendFeatureLoader({
               async *loader(__) {
                 yield createFrontendPlugin({
-                  id: 'plugin-1',
+                  pluginId: 'plugin-1',
                   extensions: [
                     createExtension({
                       name: '1',
@@ -202,7 +202,7 @@ describe('createFrontendFeatureLoader', () => {
                 yield createFrontendFeatureLoader({
                   loader: async ___ => [
                     createFrontendPlugin({
-                      id: 'plugin-2',
+                      pluginId: 'plugin-2',
                       extensions: [
                         createExtension({
                           name: '2',
@@ -222,7 +222,7 @@ describe('createFrontendFeatureLoader', () => {
               },
             }),
             createFrontendPlugin({
-              id: 'plugin-output',
+              pluginId: 'plugin-output',
               extensions: [
                 createExtension({
                   name: 'output',
@@ -278,7 +278,7 @@ describe('createFrontendFeatureLoader', () => {
           [
             nestedFeatureLoaderHolder.loader,
             createFrontendPlugin({
-              id: 'plugin',
+              pluginId: 'plugin',
               extensions: [
                 createExtension({
                   name: 'output',
@@ -319,7 +319,7 @@ describe('createFrontendFeatureLoader', () => {
 
   it('should support multiple output formats', async () => {
     const feature = createFrontendPlugin({
-      id: 'test',
+      pluginId: 'test',
     });
     const dynamicFeature = Promise.resolve({ default: feature });
 
@@ -411,7 +411,7 @@ describe('createFrontendFeatureLoader', () => {
 
   it('should limit feature loading recursion', async () => {
     const plugin = createFrontendPlugin({
-      id: 'plugin',
+      pluginId: 'plugin',
       extensions: [
         createExtension({
           name: 'output',