+++ plugins/auth-backend-module-oidc-provider/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-auth-backend-module-oidc-provider
 
+## 0.4.3-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-auth-backend@0.25.0-next.1
+  - @backstage/plugin-auth-node@0.6.3-next.1
+  - @backstage/backend-plugin-api@1.3.1-next.1
+  - @backstage/config@1.3.2
+  - @backstage/types@1.2.1
+
 ## 0.4.3-next.0
 
 ### Patch Changes