+++ plugins/kubernetes-common/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-kubernetes-common
 
+## 0.9.5-next.0
+
+### Patch Changes
+
+- 216c6b2: Updated dependency `@kubernetes/client-node` to `1.1.2`.
+- 72d019d: Removed various typos
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.9.0-next.0
+  - @backstage/catalog-model@1.7.3
+  - @backstage/types@1.2.1
+
 ## 0.9.4
 
 ### Patch Changes