+++ plugins/techdocs-module-addons-contrib/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-techdocs-module-addons-contrib
 
+## 1.1.24-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.17.2-next.0
+  - @backstage/frontend-plugin-api@0.10.2-next.0
+  - @backstage/integration@1.16.4-next.1
+  - @backstage/plugin-techdocs-react@1.2.17-next.0
+  - @backstage/integration-react@1.2.7-next.1
+  - @backstage/core-plugin-api@1.10.6
+
 ## 1.1.24-next.0
 
 ### Patch Changes